pyproject.toml
[tool.poetry]
name = "poet-glp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "poet_glp"}]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^5.30.0"
pandas = "^1.4.4"
requests = "^2.28.1"
mypy = "^0.971"
APScheduler = "^3.9.1"
mysqlclient = "^2.1.1"
jupyter = "^1.0.0"
ccxt = "^1.95.11"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
